Updated the way scales are formed. Before scales were hardcoded, meaning every scale had to be put in a dictionary. Therefore an entry had to be made for the C major, D major, E major, and so on. Since the intervals (which following notes to kind of 'skip') are the same for every major scale, I've updated the calculation of the scale. The only entry for the major scale is the interval values. Therefore the function calculates the scale based on the interval instead of hardcoded values.
Next up is the way the note dictionary works. At the moment there is a risk of an IndexError, since the starting point is the root note and the intervals just get added to the index of the root note. Meaning if the index is (for example) in the middle of the dictionary, the intervals add up to outside of the index range. This is simply countered by two ranges of notes, since the built in index-function starts at the first encounter of the value passed in. This way whenever the addition gets out of the note range (A - G#), it just enter another range of notes (see note_dict in scale_db.py). What it needs to do is simply start at the beginning of the range.